+++ designer/client/src/components/common/dndItems/FakeFormWindow.tsx
@@ -1,6 +1,6 @@
 // provide parents css classes to dragged clone - temporary.
 import React, { PropsWithChildren } from "react";
-import { NodeTable } from "../NodeDetailsContent/NodeTable";
+import { NodeTable } from "../../graph/node-modal/NodeDetailsContent/NodeTable";
 
 export function FakeFormWindow({ children }: PropsWithChildren<unknown>): JSX.Element {
     return (
