+++ designer/client/src/components/toolbarComponents/toolbarWrapper/ToolbarWrapper.tsx
@@ -4,7 +4,7 @@ import { toggleToolbar } from "../../../actions/nk/toolbars";
 import { getIsCollapsed, getToolbarsConfigId } from "../../../reducers/selectors/toolbars";
 import ErrorBoundary from "../../common/ErrorBoundary";
 import { variables } from "../../../stylesheets/variables";
-import { useDragHandler } from "../DragHandle";
+import { useDragHandler } from "../../common/dndItems/DragHandle";
 import { CollapsiblePanelContent, Panel, PanelHeader } from "../Panel";
 import { IconWrapper, StyledCloseIcon, StyledCollapseIcon, Title } from "./ToolbarStyled";
 import { useTheme } from "@mui/material";
