+++ designer/client/src/components/toolbars/Survey.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from "react-i18next";
 import React, { useCallback } from "react";
 import { ToolbarWrapper } from "../toolbarComponents/toolbarWrapper/ToolbarWrapper";
-import { DragHandle } from "../toolbarComponents/DragHandle";
+import { SimpleDragHandle } from "../common/dndItems/DragHandle";
 import { Button, Stack, Typography, useTheme } from "@mui/material";
 import { useWindows, WindowKind } from "../../windowManager";
 import { useSurvey } from "./useSurvey";
@@ -32,7 +32,7 @@ function Survey(): JSX.Element {
 
     return (
         <ToolbarWrapper onClose={hideSurvey} color={theme.custom.colors.accent}>
-            <DragHandle>
+            <SimpleDragHandle>
                 <Stack p={1} spacing={0.5}>
                     <Typography variant="body2">{survey.text}</Typography>
                     <Stack direction="row" spacing={1}>
@@ -44,7 +44,7 @@ function Survey(): JSX.Element {
                         </Button>
                     </Stack>
                 </Stack>
-            </DragHandle>
+            </SimpleDragHandle>
         </ToolbarWrapper>
     );
 }
